feat: rename `dps-autodate` to `autoreplacer.vim`
This commit renames the `dps-autodate` plugin to `autoreplacer` to better reflect its functionality.

The old `dps-autodate` plugin was designed to automatically set dates in documents, while the new `autoreplacer` plugin is designed to automatically replace any specified keyword with a new value. This change will make it easier for users to find and use the plugin, and will also make it more consistent with other plugins in the ecosystem.